+++ docs/changelog.md
@@ -1,4 +1,6 @@
 - Add the validator version to the validation OperationOutcome, as per the FHIR Tooling Extensions (#581)
+- Add proposed profiles in the validator even if a `meta.profile` is present (#335)
+- Add information about the proposed profiles in the validator: package ID and version, title, and canonical URL (#355)
 
 2026/09/14 Release 4.1.16